Split invoices package into invoices and invoices/invutil
invoices contains core database functions and invutil contains
convenience utility functions. Reason was an otherwise upcoming
cyclic reference while creating the new Invoice for a new
activation.

+++ controllers/billing/migrate_ctrl.go
@@ -159,7 +159,7 @@ func (this *Controller) Migrate() {
 		case 1:
 			if p.InvoiceId == 0 {
 				inv := invs[0]
-				p.InvoiceId = uint64(inv.Id)
+				p.InvoiceId = inv.Id
 			}
 		default:
 			beego.Error("Matched", len(invs), "invoices to purchase", p.Id)
